+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Restrict Asset Delete Changelog
 
+## 1.1.0 - 2018-11-22
+### Added
+- Ability to set the restriction permission per user
+- Ability to skip the restriction for Admin users
+
 ## 1.0.0 - 2018-10-24
 ### Added
 - Initial release